Update kernel error types
