+++ src/kernels/errors/jupyterKernelDependencyError.ts
@@ -6,7 +6,7 @@
 import { DataScience } from '../../platform/common/utils/localize';
 import { getDisplayNameOrNameOfKernelConnection } from '../helpers';
 import { KernelConnectionMetadata, KernelInterpreterDependencyResponse } from '../types';
-import { BaseKernelError } from '../../platform/errors/types';
+import { BaseKernelError } from './types';
 
 export class JupyterKernelDependencyError extends BaseKernelError {
     constructor(
