+++ src/kernels/errors/kernelInterruptTimeoutError.ts
@@ -3,7 +3,7 @@
 
 import { DataScience } from '../../platform/common/utils/localize';
 import { KernelConnectionMetadata } from '../types';
-import { BaseKernelError } from '../../platform/errors/types';
+import { BaseKernelError } from './types';
 
 export class KernelInterruptTimeoutError extends BaseKernelError {
     constructor(kernelConnection: KernelConnectionMetadata) {
