+++ src/notebooks/controllers/kernelConnector.ts
@@ -27,8 +27,8 @@ import { Product } from '../../kernels/installer/types';
 import { INotebookControllerManager, INotebookEditorProvider } from '../types';
 import { selectKernel } from './kernelSelector';
 import { KernelDeadError } from '../../kernels/errors/kernelDeadError';
+import { IDataScienceErrorHandler } from '../../kernels/errors/types';
 import { noop } from '../../platform/common/utils/misc';
-import { IDataScienceErrorHandler } from '../../platform/errors/types';
 import { IStatusProvider } from '../../platform/progress/types';
 import { IRawNotebookProvider } from '../../kernels/raw/types';
 import { IVSCodeNotebookController } from './types';
