+++ src/platform/errors/types.ts
@@ -1,29 +1,10 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
 
-import {
-    KernelAction,
-    KernelActionSource,
-    KernelConnectionMetadata,
-    KernelInterpreterDependencyResponse
-} from '../../kernels/types';
-import { Resource } from '../common/types';
-
 export abstract class BaseError extends Error {
     public stdErr?: string;
     constructor(public readonly category: ErrorCategory, message: string) {
         super(message);
-    }
-}
-
-export abstract class BaseKernelError extends BaseError {
-    public override stdErr?: string;
-    constructor(
-        category: ErrorCategory,
-        message: string,
-        public readonly kernelConnectionMetadata: KernelConnectionMetadata
-    ) {
-        super(category, message);
     }
 }
 
@@ -59,15 +40,6 @@ export class WrappedError extends BaseError {
             err = err.originalException;
         }
         return err;
-    }
-}
-export class WrappedKernelError extends WrappedError {
-    constructor(
-        message: string,
-        originalException: Error | undefined,
-        public readonly kernelConnectionMetadata: KernelConnectionMetadata
-    ) {
-        super(message, originalException);
     }
 }
 
@@ -134,28 +106,3 @@ export type TelemetryErrorProperties = {
      */
     pythonErrorPackage?: string;
 };
-
-export const IDataScienceErrorHandler = Symbol('IDataScienceErrorHandler');
-export interface IDataScienceErrorHandler {
-    /**
-     * Handles the errors and if necessary displays an error message.
-     */
-    handleError(err: Error): Promise<void>;
-    /**
-     * Handles errors specific to kernels.
-     * The value of `errorContext` is used to determine the context of the error message, whether it applies to starting or interrupting kernels or the like.
-     * Thus based on the context the error message would be different.
-     */
-    handleKernelError(
-        err: Error,
-        errorContext: KernelAction,
-        kernelConnection: KernelConnectionMetadata,
-        resource: Resource,
-        actionSource: KernelActionSource
-    ): Promise<KernelInterpreterDependencyResponse>;
-    /**
-     * The value of `errorContext` is used to determine the context of the error message, whether it applies to starting or interrupting kernels or the like.
-     * Thus based on the context the error message would be different.
-     */
-    getErrorMessageForDisplayInCell(err: Error, errorContext: KernelAction): Promise<string>;
-}
