+++ src/webviews/extension-side/dataviewer/dataViewer.ts
@@ -39,12 +39,12 @@ import {
 } from '../../../platform/common/types';
 import * as localize from '../../../platform/common/utils/localize';
 import { StopWatch } from '../../../platform/common/utils/stopWatch';
-import { IDataScienceErrorHandler } from '../../../platform/errors/types';
 import { Telemetry } from '../../webview-side/common/constants';
 import { WebViewViewChangeEventArgs } from '../types';
 import { WebviewPanelHost } from '../webviewPanelHost';
 import { noop } from '../../../platform/common/utils/misc';
 import { joinPath } from '../../../platform/vscode-path/resources';
+import { IDataScienceErrorHandler } from '../../../kernels/errors/types';
 
 const PREFERRED_VIEWGROUP = 'JupyterDataViewerPreferredViewColumn';
 @injectable()
